Fixed bug with rolling initiative, added tie breaker
When rolling initiative, a > sign was used to determine which creature goes first. This lead to a ~2.5% favor towards the first creature regardless of it's stats. This was fixed by adding a tie breaker. Now in the event of a tie, each creature has a 50% chance of getting picked to go firsts.